move to hook for ease of testing
